refactor(Movie): 패키지 구조 변경으로 인한 import문구 수정됨
